feat: Add function to check entry object in deque
Adds `check_entry_object_in_deque` function which measures the performance of checking for the existence of a generated UUID string within a deque. This function runs a given number of checks, generating a new UUID string in each iteration and checking if it exists within the provided deque. A timer decorator is used to log the function's execution time. The deque data structure is imported from the collections module.