feat: Add function to check entry object in dictionary
This commit introduces a new function, `check_entry_object_in_dict`, which checks if a randomly generated UUID (as a hex string) exists as a key in a given dictionary. This function is decorated with `@timer_of_execution` to measure its performance. The function iterates `number_of_checks` times, generating a new UUID for each iteration. The primary use-case is for benchmarking the performance of dictionary lookups with varying amounts of data and under different operational parameters.