feat: Allow generating multiple copies of test data
This change introduces a new `num_of_objects` parameter to the `get_test_data_from_json` function, allowing the user to specify how many copies of the data objects from the JSON file should be yielded.

This is useful when needing to generate a larger dataset from a smaller JSON test data file. It uses deepcopy to avoid modifying the original data.