refactor: Refactor benchmark_id_generation function
This commit refactors the benchmark_id_generation function to accept a list of iterators instead of a single number of IDs. This change allows for more flexibility in the benchmarking process, as different iterators can be used to generate IDs. The measure_resource_usage function now accepts a redis_iterator and a callable function.
The memory_object_report calls were moved to the end of the function.