refactor: Improve membership benchmark with iterator input
This commit refactors the membership benchmark to accept iterators as input for generating the data collections. This change allows the benchmark to be more flexible and efficient when dealing with large datasets, as it avoids loading all the data into memory at once.

The `run_membership_benchmark` function now takes a list of iterators and a list of keys as input. It generates the list, deque, set, and dict collections from these iterators. The `analyze_membership_performance` function is also updated to use the list of keys for membership tests.