feat: Add CPU monitoring decorator with configurable options
This commit introduces a decorator, `cpu_monitor_decorator`, that enables CPU usage monitoring for decorated functions.  The decorator allows for optional plotting of CPU usage, printing to the console, and saving data to a file. A flag `is_enabled` is added to allow easily enable/disable decorator.
The monitor run in separate thread using `psutil` library.