feat: Add Redis support for storing CPU and RAM usage data
This change introduces the ability to store CPU and RAM usage data in Redis, providing a more persistent and accessible storage solution.
The `ram_monitor_decorator` and `cpu_monitor_decorator` functions have been updated to accept an optional `r_client` argument, which is a Redis client instance. If provided, the usage data will be stored in Redis using the function name as the key. Also, the data saving mechanism is updated to use json dumps/loads.
This enhancement allows for easier analysis and visualization of performance data, especially in distributed systems.