feat: Add script to generate test data for Redis benchmark
This change adds a `if __name__ == "__main__":` block to `redis_test_data_loader.py` which will generate a file called `test_data.json` when the file is executed.

This allows us to easily generate test data with a specified number of
keys and a specified number of keys with actual data, for Redis
benchmarking purposes.