refactor: Rename redis_client to r_client_prepare
This change renames the `redis_client` variable to `r_client_prepare` to improve code clarity and reflect its purpose in preparing the Redis database for tests. This change affects all functions that used the client. Also, add list in redis with keys which contain some data for faster usage in benchmarks.