feat: Refactor benchmark to accept collection type
This commit refactors the `run_benchmark_simple_collections` function to accept the collection type as an argument, rather than a numerical index. This makes the benchmark more flexible and easier to use, as the user can now specify the collection type directly. It also adds a parameter `is_print_size` that enables or disables printing the size of the created collection.