feat: Refactor benchmark to accept collection type
This commit refactors the `run_benchmark_create_collections_real_data` function to accept the collection type as an argument instead of an integer. This allows for more flexibility in running benchmarks for different collection types.  It also adds an optional `is_print_size` parameter to control whether the size of the collection is printed.