feat: Add scenario_v7 to insert data at the beginning of list
This commit introduces `scenario_v7` to the benchmark suite.
This new scenario focuses on measuring the performance impact of inserting data at the beginning of a list, which can be less efficient than appending to the end. The scenario uses `raw_data_list.insert(0, ...)` to add elements to the beginning of the list.