refactor: Allow passing redis_iterator to run_benchmark function
This change allows the `redis_iterator` to be passed directly to the `run_benchmark_simple_collections` function. Previously, the iterator was hardcoded within the `create_collection_with_simple_data` function with a default value. By making it a parameter of `run_benchmark_simple_collections`, we gain more control over the benchmark's data source, improving flexibility and testability. The iterator is now explicitly passed when calling the benchmark in the main block.